Componentize new task form
